Add another test
Converted chapel.inline.multi.py to test_chapel_add_assigns.py.
